fix change rate of restaurant

+++ Backend/dashboard/models.py
@@ -9,7 +9,7 @@ class siteAdminModel(models.Model):
     restaurantType = models.CharField(max_length=50)
     restaurantImage = models.FileField(upload_to='restaurantImage/')
     restaurantLocation = models.CharField(max_length=50)
-    restaurantRate = models.IntegerField()
+    restaurantRate = models.FloatField()
     restaurantRateNumber = models.IntegerField()
     restaurantUsername = models.CharField(max_length=50)
     restaurantPassword = models.CharField(max_length=50)
